+++ src/rekordbox/mod.rs
@@ -20,7 +20,6 @@ mod packets;
 mod db_field;
 mod db_request_type;
 mod db_message_argument;
-mod metadata_type;
 mod rpc;
 mod status_event_server;
 mod keepalive;
@@ -41,3 +40,4 @@ use library::DBLibraryServer;
 pub use packets::DBMessage;
 pub use library::{Track, Metadata};
 pub use library::database::Database;
+use library::metadata_type;
